moved highlight to util module
